add tint color support
