+++ README.md
@@ -4,7 +4,7 @@
 [![pub package](https://img.shields.io/pub/v/soft_edge_blur.svg)](https://pub.dev/packages/soft_edge_blur)
 
 
-A Flutter package that provides a customizable soft edge blur effect for widgets.
+A Flutter package that provides a customizable soft progressive blur effect for widgets.
 
 <img width="547" alt="Map blurred" src="https://i.imgur.com/7DcixDz.png">
 
@@ -34,20 +34,26 @@ import 'package:soft_edge_blur/soft_edge_blur.dart';
 Wrap any widget with `SoftEdgeBlur` to apply the blur effect:
 
 ```dart
-SoftEdgeBlur(
+return SoftEdgeBlur(
   edges: [
     EdgeBlur(
-      EdgeType.topEdge,
-      100.0, // Edge Size
-      20.0, // Blur Sigma
+      type: EdgeType.topEdge,
+      size: 1000,
+      sigma: 30,
       controlPoints: [
-        ControlPoint(position: 0.5, type: ControlPointType.visible),
-        ControlPoint(position: 1.0, type: ControlPointType.transparent),
+        ControlPoint(
+          position: 0.5,
+          type: ControlPointType.visible,
+        ),
+        ControlPoint(
+          position: 1,
+          type: ControlPointType.transparent,
+        )
       ],
-    ),
+    )
   ],
   child: YourWidget(),
-)
+);
 ```
 
 ## Customization
@@ -66,34 +72,19 @@ Specify which edges to apply the blur effect:
 You can apply blur to multiple edges simultaneously.
 
 ### Edge Size
-
 Set the size of the blurred area. This determines how far the blur effect extends from the edge of the widget.
 
 ### Blur Sigma
-
 Adjust the intensity of the blur effect.
 
-### Control Points
+### Tint Color
+Apply a tint color on top of the blurred area.
 
+### Control Points
 Define points to control the blur gradient along the edge. Each control point has two properties:
 
 - `position`: A value between 0.0 and 1.0, representing the position along the edge.
 - `type`: Either `ControlPointType.visible` or `ControlPointType.transparent`.
-
-
-### Example Configuration
-
-```dart
-EdgeBlur(
-  EdgeType.topEdge,
-  100.0, // Edge Size
-  20.0, // Blur Sigma
-  controlPoints: [
-    ControlPoint(position: 0.5, type: ControlPointType.visible),
-    ControlPoint(position: 1.0, type: ControlPointType.transparent),
-  ],
-),
-```
 
 ### Try it live here
