+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: soft_edge_blur
-description: "A customizable Flutter widget that applies a soft blur effect to one or more edges of its child widget."
-version: 0.1.0
+description: "A customizable Flutter widget that applies a soft progressive blur effect to one or more edges of its child widget."
+version: 0.1.1
 repository: https://github.com/Rahiche/soft_edge_blur
 
 environment:
